refactor: added style calculating to helpers
