+++ src/app/components/Assignments/AssignmentItem/AssignmentItem.tsx
@@ -4,6 +4,7 @@ import AssignmentItemBody from '../AssignmentItemBody/AssignmentItemBody';
 import AssignmentItemForm, { AssignmentFormData } from '../AssignmentItemForm/AssignmentItemForm';
 import { INormalizeAssignment } from 'core/reducers/assignments';
 import { classNames } from 'core/styles';
+import { setCardStyle } from 'core/helpers';
 
 const cn = classNames(require('../index.scss'));
 
@@ -30,32 +31,7 @@ class AssignmentItem extends React.PureComponent<AssignmentItemProps> {
     render() {
         const { isEndAssignment } = this.props.assignment;
         const { status, title, urlToDescription, score, _id } = this.props.assignment.assignment;
-        const card = {
-            className: '',
-            cardHeader: {
-                title: '',
-                className: '',
-            },
-        };
-        if (isEndAssignment) {
-            card.className = 'border-danger';
-            card.cardHeader.title = 'The deadline has passed!';
-            card.cardHeader.className = 'text-white bg-danger';
-        } else if (status === 'Assigned') {
-            card.className = 'bg-secondary';
-            card.cardHeader.title = 'Not submitted yet!';
-            card.cardHeader.className = '';
-        } else if (status === 'Checked') {
-            if (score === 100) {
-                card.className = 'border-success';
-                card.cardHeader.title = 'Done!';
-                card.cardHeader.className = 'text-white bg-success';
-            } else {
-                card.className = 'border-warning';
-                card.cardHeader.title = 'Done!';
-                card.cardHeader.className = 'text-white bg-warning';
-            }
-        }
+        const card = setCardStyle(isEndAssignment, score, status);
 
         return (
             <Card className={`bm-3 mb-4 flex-grow-0 flex-shrink-1 ${cn('card')}  ${card.className}`}>
