+++ src/app/core/helpers/index.ts
@@ -1,1 +1,2 @@
 export * from './schedule';
+export * from './assignment';
